rename: content 도메인의 repository 파일들을 /repository로 이동
