+++ mopl-api/src/main/java/io/mopl/api/content/repository/ContentQueryRepository.java
@@ -1,4 +1,4 @@
-package io.mopl.api.content.domain;
+package io.mopl.api.content.repository;
 
 import io.mopl.api.content.dto.ContentPage;
 import io.mopl.api.content.dto.ContentSearchRequest;
